moves surface pages into their own folder to avoid confusion in future development
